add Cargo.lock to repository
For wallet software we don't want to allow random updates of dependencies,
though there shouldn't be any security risk here anyway given that the
Ledger itself is a security boundary.

See also https://github.com/servo/unicode-bidi/issues/46 , it seems like
semver is not respected even for flagship Rust products so it's probably
best not to rely on it anyway.