clippy: use `write_all` rather than `write` and fix `write`
The existing code does a strange thing where `CryptWriter::write`
contains an internal implementation of `Write::write_all`. This
is weird and unnecessary.

The code then calls `cryptwriter.write()`, which is fine in this case,
but Clippy interprets it as a bug because in general, you should always
use `write_all`.