update miniscript to 10, rust-bitcioin to 0.30
The rust-bitcoin 0.29 -> 0.30 update was the "update from hell" which
renamed a ton of types and modules, dropped hex decoding (which we
fortunately don't need), renamed Script to ScriptBuf, and introduced the
`NetworkChecked` attribute for the `Address` type (which we also don't
need because we never parse addresses, only generate them).

This commit just fixes all the compilation errors. It does not do any
cleanups. We'll clean some things up after updating all the way to
Miniscript 12.